:rewind: move assets back into src/
